[curry/branches/type-classes@2049]:
The compiler now systematically renames method identifiers in the left
hand sides of method implementations in class and instance
declarations. This nicely reflects the fact that a method identifier
in the left hand side of a method declaration denotes a particular
implementation, whereas an occurrence in the right hand side denotes
the overloaded type class method.

* Renaming.lhs(renameTopDecl, renameMethodDecl),
  PrecCheck.lhs(checkMethodDecl),
  TypeCheck.lhs(tcClassMethodDecl, genClassMethodDecl, genInstMethodDecl,
                tcMethod),
  DictTrans.lhs(defaultMethodDecls, defaultMethodDecl, instDecl,
                orderMethodDecls, dictExpr)