[curry/branches/type-classes@2096]:
Bug fix: The compiler now correctly reports an error for existentially
quantified type variables that escape their scope only through the
context of a function's type. This is possible if the type variable
occurs in the context as an argument of a constrained type variable.
E.g., for
  data Key a = forall b . Key b (b -> a)
  bad (Key x f) xs = fmap (const x) xs == fmap (const x) xs
the type
  (Eq (f _c), Functor f) => Key a -> f b -> Bool
is inferred for function bad, where _c is the existentially quantified
type variable from x's type.

* TypeCheck.lhs(tcEquation, tcGoal, checkSkolem, skolemEscapingScope):
  Check for escaping skolem variables in the current context, too.