[curry/branches/type-classes@2315]:
Merge r2134 from trunk:

Introduce a wrapper function for the monadic bind primitive (>>=) in
DebugPrelude so that the debugging transformation can generate type
correct code for programs involving IO computations. The current
wrapper is not perfect, for instance, given the expression
  return (f x) >>= \y -> return (g y) -> \z -> return (h z)
(executed in an unsafePerformIO context), it will ask questions only
about the application of f, but not about g or h. Nevertheless, it is
at least sufficient to allow debugging programs that use arrays.

* DTransform.lhs(debugBind, firstPhaseQual):
  Add another special case to handle the higher-order primitive (>>=).

* curry/DebugPrelude.curry(bind'):
  Define the wrapper function for (>>=).