[curry/branches/type-classes@2321]:
Merge r2140 from trunk:

Add a check to DTransform.generateForeign that reports an error when
it is applied to an unknown higher order primitive or primitive with
IO arguments. Special wrappers (like DebugPrelude.return' and friends)
must be provided for such primitives since the standard transformation
is valid only for first order primitives with non-IO arguments.

* DTransform.lhs(generateForeign, argumentTypes, isArrowType)