[curry/branches/type-classes@2400]:
Merge r2397 from trunk:

Use a fixed naming convention for naming lambda abstractions during
desugaring, which is based on their position in the source code. With
this change it becomes possible to associate trust information with
lambda abstractions. This is used to trust a lambda abstraction if and
only if its enclosing function is trusted.

* Ident.lhs(lambdaId):
  New function returning the name of a lambda abstraction at a
  specified position.

* Desugar.lhs(desugarExpr):
  Use Ident.lambdaId for naming lambda abstractions rather than
  generating fresh names.

* Trust.lhs(instance SyntaxTree Expression):
  Associate a trust annotation with the prospective name of a lambda
  abstraction.