[curry/branches/type-classes@2434]:
The compiler no longer assumes that all free type variables of the
context of a qualified type are free in the plain type as well. In
particular, the compiler no longer reports an ambiguous type error if
the free variables of the context of an inferred type are bound in the
initial type environment. This is sound because such variables are
always monomorphic and therefore there is no ambiguity.

* Types.lhs(instance IsType QualType),
  TypeSubst.lhs(instance SubstType TypeScheme),
  TypeCheck.lhs(checkSkolems):
  The IsType QualType instance now also returns the free variables of
  the context of a qualified type.

* Types.lhs(instance IsType [a]),
  TypeCheck.lhs(fvEnv, fsEnv):
  Provide a list instance of IsType for convenience.

* TypeCheck.lhs(tcDeclGroup, dfltDecl):
  Do not report an ambiguous type error for free type variables of the
  context that are bound in the initial type environment.