[curry/branches/type-classes@2442]:
Remove the (somewhat dubious) hack that was substituting type
variables which are ambiguous in only some declarations of a
declaration group locally in those declarations. As a consequence some
rather esoteric programs, where some constrained type variables in a
declaration group do not appear in the types of all declarations of
that group, might now be rejected because the type checker infers less
general types. However note that this can be fixed by adding explicit
type signatures to the program.

* TypeCheck.lhs(tcDeclGroup, dfltDecl)