[curry/branches/type-classes@2516]:
Merge r2480 from trunk:

Fix an issue where the compiler could report an error for an export
specification T(C) depending on the order of import declarations. In
particular, this error occurred when the type T is imported more than
once and its constructor C is hidden in the last import declaration
importing type T.

This problem was introduced in r2509 (r2465 on trunk) when starting to
construct the identifier environments from scratch rather than
deriving them from the semantic environments.

* IdentInfo.lhs(instance Entity TypeKind):
  Introduce a non-standard implementation of the merge method, which
  computes the union of the constructor lists associated with two data
  constructors.

* TypeInfo.lhs(instance Entity TypeInfo),
  Imports.lhs(importUnifyData),
  Modules.lhs(importModules, importInterfaces):
  Remove the non-standard implementation of the merge method since the
  constructor list associated with a data type is used only for
  creating module interfaces where the compiler always refers to the
  original definition. This also means that the compiler no longer
  needs to merge the imports of the type constructor environment after
  loading all interfaces and importUnifyData is used only for the type
  identifier environment.