[curry/branches/type-classes@2608]:
Merge r2601 from trunk:

Simplify the build script generated by the cymk tool. The script no
longer includes explicit out of date tests. These are now implemented
by the compile and link functions defined in the cymake script.

This change also fixes the issue that "cymake -Tgoal target" did not
print the goal's type if an executable exists for target and is not
out of date. In addition, the executable is now always relinked if a
goal was specified with the -e option.

* CurryDeps.lhs(makeBuildScript),
  cymake.in(compile, link):
  The generated build script no longer includes out of date checks.

* cymk.hs(data Options, defaultOptions, data Option, options, processFiles),
  CurryDeps.lhs(buildScript, makeBuildScript, makeCleanScript),
  cymake.in(link, remove):
  The -a option is handled by the cymake script alone.

* CurryDeps.lhs(buildScript, makeBuildScript):
  Cymk's -o option now denotes the file to which the script is written
  instead of the name of the executable when generating a build script.

* cymake.1.in:
  Update the man page to reflect the change of semantics of the -e
  option and document the -a option as deprecated.

* newer.lhs(newer):
  Ignore any dependencies whose name starts with -M for compatibility
  with the link commands generated by the cymk tool.