[curry/branches/type-classes@2613]:
Merge r2606 from trunk:

Setting the environment variable HC is now the preferred way to select
a particular Haskell compiler. The option --with-hc is deprecated and
will go away soon. For now it just prints a warning when it is
used. In addition, the code for the automatic detection of a suitable
Haskell 98 compiler has been streamlined a bit.

* configure.in,
  aclocal.m4(CURRY_PROG_GHC, CURRY_PROG_HBC, CURRY_PROG_NHC):
  These macros now set HC if their respective compiler supports
  Haskell 98 and they no longer determine the compiler version.
  Furthermore, CURRY_PROG_NHC now checks only for the command name
  nhc98.

* configure.in,
  aclocal.m4(CURRY_HC_VERSION, CURRY_GHC_IOEXTS):
  The Haskell 98 compiler command is now always bound to the shell
  variable HC.

* aclocal.m4(CURRY_HC_VERSION):
  The compiler version string is now cached.