[curry/branches/type-classes@2640]:
Merge r2635 from trunk:

The default sizes of the heap, the stack, and the trail, respectively,
used in the runtime system are now defined in terms of the machine's
word size, thereby acknowledging the fact that the amount of used
memory depends on that value. The new defaults are chosen such that
the same defaults as before are used on 32-bit architectures.

* runtime/main.nw(DEFAULT_HEAPSIZE, DEFAULT_STACKSIZE, DEFAULT_TRAILSIZE):
  Change these macros into 1*M*word_size and 128*k*word_size,
  respectively.

* cyc.1.in:
  Update the man page to reflect the larger default sizes on 64-bit
  targets.