[curry/branches/type-classes@2641]:
Merge r2636 from trunk:

The default heap size of a program specified with the compiler's -h
option as well as its actual size specified with the -h runtime system
option now denote the total size of the heap rather than the size of
one of its semi-spaces. Thus, these sizes describe the actual memory
requirements of a program more accurately. The (total) default size of
the heap specified in the runtime system is not changed.

* runtime/gc_2space.nw(init_heap):
  Use the specified size directly to determine size of the heap.

* runtime/main.nw(DEFAULT_HEAPSIZE):
  Double the value so that the actual default heap size remains
  unchanged.

* cyc.1.in:
  Update the man page to reflect the new meaning of the -h options.