[curry/branches/type-classes@2898]:
Merge r2891 from trunk:

Fix the expansion of module M in the export list of a module to fully
comply with the Haskell'98 report. The old implementation was
exporting too many entities when a data constructor or field label is
in scope with a qualified name only, but its type is in scope with an
unqualified and a qualified name. On the other hand, too few entities
were exported when the unqualified name of an entity is brought into
scope by an unqualified import of a different module M'.

* TopEnv.lhs(unqualBindings, localBinding, moduleBindings),
  Interfaces.lhs(qualifyEnv1):
  The function moduleBindings, which replaces the former
  moduleImports, now returns exactly the list of entities which are in
  scope with an unqualified name x and a qualified name M.x without
  reference to the module from which the entity was imported.

* ExportSyntaxCheck.lhs(expandSpecs, expandExport):
  Remove the special case for module M in an export list where M is
  equal to the current module, as it no longer requires special
  treatment.

* ExportSyntaxCheck.lhs(expandModule):
  Restrict the list of data constructors and field labels exported
  together with a type to those which are in scope with an unqualified
  and a qualified name.

* ExportSyntaxCheck.lhs(expandLocalModule):
  Remove a redundant test for top-level definitions.

* Ident.lhs:
  Fix an obsolete comment about the assumptions made by TopEnv.