[curry/branches/type-classes@2928]:
Merge r2924 from trunk:

Fix bug where a unification x=:=g in an encapsulated search did not
force evaluation of the global variable g's value to normal form. If
the value is not evaluated elsewhere, the bug could manifest in
non-normal forms being shown as result of a goal in the interpreter.

* runtime/unify.nw(occurs, unify_var, unify_var_2):
  The occurs check no longer returns occurs_no_nf when the node is a
  global variable whose value contains unevaluated applications and
  unify_var sets up a return continuation to evaluate such variables
  to normal form.