[curry/branches/type-classes@2981]:
Merge r2980 from trunk:

Systematically rename local variables that are introduced by the
compiler. This fixes an issue where the compiler would fail with an
internal error when abstracting a local variable declaration that was
introduced by the simplifier and then eta-expanded. In addition, this
ensures that the eta-expanded function is not exported.

* Desugar.lhs(freshVar),
  Records.lhs(freshVar),
  Newtype.lhs(freshVar),
  LazyPatterns.lhs(freshVar),
  CaseMatch.lhs(freshVar),
  Simplify.lhs(freshVar),
  PatternBind.lhs(freshVar),
  DictTrans.lhs(freshVar)