[curry/branches/type-classes@3006]:
Merge r2993 from trunk:

Fix an embarrassingly long-standing bug (introduced in r1922 between
releases 0.9.10 and 0.9.11!) where the compiler could generate code
that violates the invariant that no function returns an indirection
node. This bug would manifest in unevaluated subterms in the computed
solution of goals and could even lead to missing solutions. This would
happen, e.g., for the goal
  let f xs = case xs of { _:_ -> xs } in [1..5] =:= f [1..5]
Note that the goal succeeded as expected when the arguments of the
unification are reversed.

* ILCompile.lhs(unaliasStmt):
  Recognize x as an alias of y after evaluating y <- eval x.