[curry/branches/type-classes@3128]:
Merge r3125 from trunk:

Change type inference order for list comprehension generators and do
statements of the form t<-e. The compiler now infers a type for the
expression e first and then checks that the type of the pattern t is
compatible with the element type and monadic result type of the
expression, respectively.

* TypeCheck.lhs(tcQual, tcStmt)