[curry/branches/type-classes@3244]:
Merge r3180 from trunk:

Change shell commands in Makefile rules to exit with the intended
status on Debian systems as well. In particular, a compile error in
the generated C code for the standard library will now cause the build
process to fail rather than leaving a seemingly successful but
incomplete build behind.

* Makefile.in,
  curry/Makefile.in: