[curry/branches/type-classes@3247]:
Merge r3193 from trunk:

Improve rebuilding of Makefiles when the configure script was updated.

In particular, when config.status is out of date run config.status
without parameters after config.status --recheck and recompute the
dependencies only in the main directory when Makefile is out of date.
In addition, fix dependencies for the makedoc script.

* Makefile.in