[curry/branches/type-classes@3289]:
Merge r3285 from trunk:

Remove test for prefix _debug# in the function for detecting operator
names in the runtime system. The test was incorrect, causing wrong
results for a few qualified names including an underscore in a
particular position, and furthemore it was redundant because since
r1381 the compiler strips that prefix from names in node infos.

* runtime/heap.nw(is_operator)