[curry/branches/type-classes@3290]:
Merge r3286 from trunk:

The detection of operator names in the runtime system is now based on
(the first character of) the last component of a name. This looks more
reasonable for local functions defined inside an operator definition.
For instance, M.<+>.f is no longer considered an operator symbol but a
plain identifier instead.

* runtime/heap.nw(is_operator)