[curry/branches/type-classes@3324]:
Record left and right hand side contexts separately for data constructors.

This is necessary to recover both contexts when there is some overlap
between their type constraints, e.g.,
  data Eq a => T a = Ord a => C a
Here, the Eq a constraint on the left hand side is implied by the
Ord a constraint on the right hand side. Therefore, C's type is
  Ord a => a -> T a
and left hand side context cannot be reconstructed from this type and
the right hand side context alone.

* ValueInfo.lhs(data ConstrInfo, conType),
  TypeTrans.lhs(toConstrType, expandConstrType, labelTypes),
  TypeCheck.lhs(skol),
  Exports.lhs(constrDecl),
  DictTrans.lhs(bindClassDict, dictTransValues, constrTypeRhs):
  Include the left hand side context in the additional data
  constructor information.

* test/typecheck/all.T, test/typecheck/tc043.curry,
  test/typecheck/tc043-modules/A.curry,  test/typecheck/tc043.type-incr.out,
  test/typecheck/tc043.type-tautology.out:
  New test cases.